[refactor] module 작업
