+++ login.html
@@ -9,7 +9,6 @@
 </head>
 <body>
     <div id="container">
-
             <a href="index.html">
                 <img src="img/login_logo.svg" alt="Logo" class = "main_logo">
             </a>
@@ -27,7 +26,7 @@
                     <img src="img/btn_visibility_off.svg" alt="Toggle Password Visibility">
                 </div>
         </div>
-                <button onclick="findUser()" type="button" id="form_button">로그인</button>
+                <button type="button" id="form_button">로그인</button>
             </form>
             <div class="login_options">
                     <p>간편 로그인 하기</p>
@@ -53,6 +52,6 @@
                     <button class="modal_btn" type="button">확인</button>
                 </div>
                 </div>
-    <script src="main.js"></script>
+    <script src="inputAuth.js" type = "module"></script>
 </body>
 </html>
