+++ signup.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <div id="container">
-            <a href="landing.html">
+            <a href="index.html">
                 <img src="img/login_logo.svg" alt="Logo" class = "main_logo">
             </a>
         <form>
@@ -39,7 +39,7 @@
             </div>
             <p class="error_msg"></p>
             </div>
-            <button onclick="findUser()" type="button" id="form_button">회원가입</button>
+            <button type="button" id="form_button">회원가입</button>
         </form>
             <div class="login_options">
                 <p>간편 로그인 하기</p>
@@ -67,6 +67,6 @@
 
             </div>
     </div>
-    <script src="index.js"></script>
+    <script src="inputAuth.js" type ="module"></script>
 </body>
 </html>
