+++ style.css
@@ -338,7 +338,7 @@ label{
     display: flex;
     justify-content: center;
     align-items: center;
-        background-color: var(--Secondary-400);
+    background-color: var(--Secondary-400);
 
     width: 100%;
     height: 3.5rem;
@@ -416,8 +416,8 @@ label{
 /*/////////////////Alert design ////////////////////*/
 
 .modal_container {
-    position: absolute;
     display: none;
+    position: absolute;
     flex-flow: column wrap;
     justify-content: center;
     width: 34rem;
